Add files via upload
i just threw some random crap together, and a lot more work needs to be done. This (https://github.com/FastLED/FastLED/wiki/Basic-usage) is helpful. Also, Ican't figure out how to call groups of LEDs without doing it one by one which is kind of a problem. The whole thing runs as one big loop, so push the reset button on the Arduino to start from the beginning.